OCR/google_vision/requirements.txt
google-cloud-vision>=3.4.0
PyMuPDF>=1.23.0
pathlib2>=2.3.7
requests>=2.28.0
protobuf>=4.21.0
python-dotenv>=1.0.0
mistralai>=1.0.0
